Remove cargo_metadata from auditable-serde API surface. I can expose it as a separate crate if anyone tells me they need it.
