+++ auditable-serde/Cargo.toml
@@ -13,17 +13,11 @@ all-features = true
 
 [features]
 default = []
-from_metadata = ["cargo_metadata"]
 schema = ["schemars"]
 
 [dependencies]
 serde = { version = "1", features = ["serde_derive"] }
 serde_json = "1.0.57"
 semver = { version = "1.0", features = ["serde"] }
-cargo_metadata = { version = "0.18", optional = true }
 topological-sort = "0.2.2"
 schemars = {version = "0.8.10", optional = true }
-
-[[example]]
-name = "from-metadata"
-required-features = ["from_metadata"]
